Refactor to registry submodule
